refactor: split Send File scene's protocols into separated folder and files
