+++ Pharus/Scenes/Send File/Protocols/SendFileFlow.swift
@@ -1,0 +1,12 @@
+//
+//  SendFileFlow.swift
+//  Pharus
+//
+//  Created by Victor Colen on 17/06/22.
+//
+
+import Foundation
+
+protocol SendFileFlow {
+    func showFileSentAlert()
+}
