+++ Pharus/Scenes/Send File/Protocols/SendFilePresenterProtocol.swift
@@ -1,0 +1,13 @@
+//
+//  SendFilePresenterProtocol.swift
+//  Pharus
+//
+//  Created by Victor Colen on 17/06/22.
+//
+
+import Foundation
+
+protocol SendFilePresenterProtocol {
+    func uploadFile()
+    func sendFile()
+}
